Upgrade to Swift 6.2 and drop Combine from CommonLibrary (#1607)
Combine is dangerously unsafe with actors.

Main changes:

- **Replace Combine with Partout subject streams**
- ~~Weird Combine/Dispatch crash in InMemoryProfileRepository~~ (due to
Combine actor-unsafety)
- Cope with [incompatible PacketTunnelProvider
signature](https://github.com/mullvad/mullvadvpn-app/issues/8750)
- Start AVSession from main queue

Secondary changes:

- Ack that isBeta is async
- Add "fake" Sendable in NIO-based implementations
- Sendable, Sendable, Sendable

+++ app-shared/Sources/AppLibraryMain/Views/App/LegacyAppCoordinator+ModalRoute.swift
@@ -37,7 +37,7 @@ extension LegacyAppCoordinator {
         }
     }
 
-    enum ConfirmationAction {
+    enum ConfirmationAction: Sendable {
         case deleteProfile(ABI.ProfilePreview)
     }
 }
